+++ subpackages/lib/source/odood/lib/assembly/exception.d
@@ -1,0 +1,14 @@
+module odood.lib.assembly.exception;
+
+private import std.exception;
+
+private import odood.exception;
+
+
+class OdoodAssemblyException : OdoodException
+{
+    mixin basicExceptionCtors;
+}
+
+
+